Split skill URL install helper
