+++ crates/ironclaw_host_runtime/src/first_party_tools/mod.rs
@@ -11,6 +11,7 @@ mod json;
 mod schemas;
 mod shell;
 mod skill_management;
+mod skill_url_install;
 mod time;
 
 use std::{sync::Arc, time::Instant};
